+++ src/main/java/nextstep/subway/auth/application/dto/GithubTokenRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.member.application.dto;
+package nextstep.subway.auth.application.dto;
 
 public class GithubTokenRequest {
     private String code;
