+++ src/main/java/nextstep/subway/auth/application/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.member.application.dto;
+package nextstep.subway.auth.application.dto;
 
 public class TokenRequest {
     private String email;
